+++ src/app/cart/cart.service.ts
@@ -1,18 +1,16 @@
-import { Injectable } from '@angular/core';
+import {Injectable} from '@angular/core';
 
-import { CartItem } from './cart-item.model';
+import {CartItem} from './item.model';
 
 @Injectable({
-  providedIn: 'root'
+    providedIn: 'root'
 })
 export class CartService {
 
-  constructor() { }
+    constructor() {
+    }
 
-  getItems(): Array<CartItem> {
-    return [
-        new CartItem(1, 'Bleyer Shoes', 57),
-        new CartItem(4, 'Belt', 20)
-    ];
-  }
+    getItems(): Array<CartItem> {
+        return [];
+    }
 }