+++ src/app/catalog/product/product.component.ts
@@ -1,0 +1,30 @@
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {Product} from '../product.model';
+
+@Component({
+    selector: 'app-product',
+    templateUrl: './product.component.html'
+})
+
+export class ProductComponent implements OnInit {
+    @Input() product: Product;
+
+    @Output() buyProduct: EventEmitter<Product> = new EventEmitter<Product>();
+
+    expanded: boolean;
+
+    constructor() {}
+
+    ngOnInit() {
+        this.expanded = false;
+    }
+
+    buy(): void {
+        console.log(`click handler inside product component`);
+        this.buyProduct.emit(this.product);
+    }
+
+    toggleDescription() {
+        this.expanded = !this.expanded;
+    }
+}