+++ src/app/shared/shared.module.ts
@@ -1,0 +1,16 @@
+import {NgModule} from '@angular/core';
+import {HoverStyleDirective} from './directives/hover-style.directive';
+import {ClickPreventDirective} from './directives/click-prevent.directive';
+
+@NgModule({
+    declarations: [
+        HoverStyleDirective,
+        ClickPreventDirective
+    ],
+    exports: [
+        HoverStyleDirective,
+        ClickPreventDirective
+    ]
+})
+
+export class SharedModule {}